refactor: move leader board page to /leaderboard
